ESP32 Bluetooth Classic (BT) support (#428)
* User defined ESP32 BLE Device name

Add possibility to define the device name of the ESP32 BLE device

* name BLE_DEVICE_NAME to be consistent with other BLE's

* Flushing buffer in direct mode

When running Blynk in direct mode (using Bluetooth/BLE), the connection can stop working after "Packet too big" or "Invalid header type". The cause is the loss of one or more bytes, which makes the Blynk protocol get out of sync, resulting in a connection being closed.
However, it is not necessary to stop the connection, it just needs to resync. Because the Blynk protocol misses sync bytes, it should be done in another way. In this solution, it makes use of the fact that Blynk on BT/BLE is sending its bytes in block, starting with the header. So, in case the systen is no longer synced, we just need to flush the buffer to get it in sync again. Using this fix, the stability of BLE in Blynk increases a lot. Because conn.flush() doesn't exist, we make use of the fact that conn.connect() is basically executing a buffer flush.

* Update BlynkProtocol.h

flush in direct mode instead of reconnecting

* update naar blynk lib

* Initial ESP32 BT classic support

* ESP32 BT classic example

* Update BlynkProtocol.h

* use old values ESP_PWR_LVL_N2, ESP_PWR_LVL_P7

Use of old values ESP_PWR_LVL_N2, ESP_PWR_LVL_P7 for backward compatibility, they are also available in new IDF code